backend/importer/requirements.txt
nameko==3.0.0-rc11
python-dotenv
pytest
pytest-cov
pandas
sqlmodel
sqlmodel-controller
openpyxl